Remove name validation in ExpressionLanguage#compile
Make it possible to use any names in parser so that you can compile an expression without really knowing what context will it have. It makes it easier to work with component for other components, but on the other hand it's less robust/more error prone because one of the validation stages is removed.